Seperated user and profile types
