+++ apps/frontend/src/types/auth.ts
@@ -1,4 +1,14 @@
-import { MentorshipRole, UserType } from './user';
+export type UserType = 'EMPLOYER' | 'JOB_SEEKER';
+export type MentorshipRole = 'MENTOR' | 'MENTEE';
+
+export interface User {
+  id: number;
+  username: string;
+  email: string;
+  bio: string;
+  userType: UserType;
+  mentorType: MentorshipRole;
+}
 
 export interface LoginCredentials {
   username: string;
