--- apps/frontend/src/types/user.ts
@@ -1,11 +1,0 @@
-export type UserType = 'EMPLOYER' | 'JOB_SEEKER';
-export type MentorshipRole = 'MENTOR' | 'MENTEE';
-
-export interface User {
-  id: string;
-  username: string;
-  email: string;
-  bio: string;
-  userType: UserType;
-  mentorType: MentorshipRole;
-}
